Get layer ordering working again
