+++ src/lib/browse/InterventionLayer.svelte
@@ -6,6 +6,7 @@
     isLine,
     isPoint,
     isPolygon,
+    layerId,
   } from "lib/maplibre";
   import type { FilterSpecification } from "maplibre-gl";
   import { colorInterventionsBySchema } from "schemas";
@@ -31,7 +32,7 @@
 
 <GeoJSON data={gj}>
   <CircleLayer
-    id="interventions-points"
+    {...layerId("interventions-points")}
     filter={["all", isPoint, hideWhileEditing, notEndpoint]}
     paint={{
       "circle-color": colorInterventions,
@@ -47,7 +48,7 @@
   </CircleLayer>
 
   <LineLayer
-    id="interventions-lines"
+    {...layerId("interventions-lines")}
     filter={["all", isLine, hideWhileEditing]}
     paint={{
       "line-color": colorInterventions,
@@ -62,7 +63,7 @@
     </Popup>
   </LineLayer>
   <CircleLayer
-    id="interventions-lines-endpoints"
+    {...layerId("interventions-lines-endpoints")}
     filter={["==", "endpoint", true]}
     paint={{
       "circle-radius": 0.5 * lineWidth,
@@ -76,7 +77,7 @@
   />
 
   <FillLayer
-    id="interventions-polygons"
+    {...layerId("interventions-polygons")}
     filter={["all", isPolygon, hideWhileEditing]}
     paint={{
       "fill-color": colorInterventions,
@@ -91,7 +92,7 @@
     </Popup>
   </FillLayer>
   <LineLayer
-    id="interventions-polygons-outlines"
+    {...layerId("interventions-polygons-outlines")}
     filter={["all", isPolygon, hideWhileEditing]}
     paint={{
       "line-color": colorInterventions,
