+++ src/lib/browse/layers/areas/LocalPlanningAuthorities.svelte
@@ -7,6 +7,7 @@
     publicResourceBaseUrl,
   } from "lib/common";
   import { Checkbox } from "lib/govuk";
+  import { layerId } from "lib/maplibre";
   import {
     FillLayer,
     hoverStateFilter,
@@ -18,7 +19,6 @@
 
   let name = "local_planning_authorities";
   let color = colors.local_planning_authorities;
-  let outlineLayer = `${name}-outline`;
 
   let show = false;
 
@@ -72,7 +72,7 @@
   url={`pmtiles://${publicResourceBaseUrl()}/v1/${name}.pmtiles`}
 >
   <FillLayer
-    id={name}
+    {...layerId(name)}
     sourceLayer={name}
     paint={{
       "fill-color": color,
@@ -88,7 +88,7 @@
     </Popup>
   </FillLayer>
   <LineLayer
-    id={outlineLayer}
+    {...layerId(`${name}-outline`)}
     sourceLayer={name}
     paint={{
       "line-color": color,
