+++ src/lib/browse/layers/areas/ParliamentaryConstituencies.svelte
@@ -7,6 +7,7 @@
     publicResourceBaseUrl,
   } from "lib/common";
   import { Checkbox } from "lib/govuk";
+  import { layerId } from "lib/maplibre";
   import {
     FillLayer,
     hoverStateFilter,
@@ -19,7 +20,6 @@
 
   let name = "parliamentary_constituencies";
   let color = colors.parliamentary_constituencies;
-  let outlineLayer = `${name}-outline`;
 
   let show = false;
 
@@ -59,7 +59,7 @@
   url={`pmtiles://${publicResourceBaseUrl()}/v1/${name}.pmtiles`}
 >
   <FillLayer
-    id={name}
+    {...layerId(name)}
     sourceLayer={name}
     paint={{
       "fill-color": color,
@@ -77,7 +77,7 @@
     </Popup>
   </FillLayer>
   <LineLayer
-    id={outlineLayer}
+    {...layerId(`${name}-outline`)}
     sourceLayer={name}
     paint={{
       "line-color": color,
