+++ src/lib/browse/layers/lines/CyclePaths.svelte
@@ -6,7 +6,7 @@
     publicResourceBaseUrl,
   } from "lib/common";
   import { Checkbox } from "lib/govuk";
-  import { constructMatchExpression } from "lib/maplibre";
+  import { constructMatchExpression, layerId } from "lib/maplibre";
   import {
     hoverStateFilter,
     LineLayer,
@@ -104,7 +104,7 @@
   url={`pmtiles://${publicResourceBaseUrl()}/v1/${name}.pmtiles`}
 >
   <LineLayer
-    id={name}
+    {...layerId(name)}
     sourceLayer={name}
     paint={{
       "line-color": constructMatchExpression(
