+++ src/lib/browse/layers/lines/MajorRoadNetwork.svelte
@@ -7,6 +7,7 @@
     publicResourceBaseUrl,
   } from "lib/common";
   import { Checkbox } from "lib/govuk";
+  import { layerId } from "lib/maplibre";
   import {
     hoverStateFilter,
     LineLayer,
@@ -42,7 +43,7 @@
   url={`pmtiles://${publicResourceBaseUrl()}/v1/${name}.pmtiles`}
 >
   <LineLayer
-    id={name}
+    {...layerId(name)}
     sourceLayer={name}
     paint={{
       "line-color": color,
