+++ src/lib/browse/layers/points/Crossings.svelte
@@ -7,6 +7,7 @@
     publicResourceBaseUrl,
   } from "lib/common";
   import { Checkbox } from "lib/govuk";
+  import { layerId } from "lib/maplibre";
   import {
     CircleLayer,
     VectorTileSource,
@@ -75,7 +76,7 @@
   url={`pmtiles://${publicResourceBaseUrl()}/v1/${name}.pmtiles`}
 >
   <CircleLayer
-    id={name}
+    {...layerId(name)}
     sourceLayer={name}
     paint={{
       "circle-color": color,
