+++ src/lib/browse/layers/points/CycleParking.svelte
@@ -7,6 +7,7 @@
     publicResourceBaseUrl,
   } from "lib/common";
   import { Checkbox } from "lib/govuk";
+  import { layerId } from "lib/maplibre";
   import { CircleLayer, VectorTileSource } from "svelte-maplibre";
   import { colors } from "../../colors";
   import OsmLicense from "../OsmLicense.svelte";
@@ -39,7 +40,7 @@
   url={`pmtiles://${publicResourceBaseUrl()}/v1/${name}.pmtiles`}
 >
   <CircleLayer
-    id={name}
+    {...layerId(name)}
     sourceLayer={name}
     paint={{
       "circle-color": color,
