+++ src/lib/common/LineMeasureController.svelte
@@ -4,7 +4,7 @@
   import { colors } from "colors";
   import DraggablePin from "lib/common/DraggablePin.svelte";
   import { SecondaryButton } from "lib/govuk";
-  import { emptyGeojson } from "lib/maplibre";
+  import { emptyGeojson, layerId } from "lib/maplibre";
   import { LngLat, MapMouseEvent } from "maplibre-gl";
   import { map } from "stores";
   import { onDestroy, onMount } from "svelte";
@@ -165,7 +165,7 @@
 
 <GeoJSON data={drawGj}>
   <LineLayer
-    id="measurement-line"
+    {...layerId("measurement-line")}
     paint={{ "line-color": colors.measuringLine, "line-width": 5 }}
   />
 </GeoJSON>
