+++ src/lib/maplibre/index.ts
@@ -1,3 +1,4 @@
 export * from "./utils";
 export * from "./layer_helper";
 export * from "./layer_helper_utils";
+export { layerId } from "./zorder";
